inject sc globals for homepage with fixed html - worked only for pages requested by name, not home
