+++ packages/server/routes/utils.js
@@ -492,6 +492,35 @@ const admin_config_route = ({
 };
 
 /**
+ * Send an HTML string response with injected Saltcorn globals (CSRF token,
+ * version tag, locale, page-load tag, admin flag) and normalised static-asset paths.
+ * @param {any} req
+ * @param {any} res
+ * @param {string} html_string
+ * @returns
+ */
+const sendHtmlStringWithGlobals = (req, res, html_string) => {
+  res.set("Content-Type", "text/html");
+  const version_tag = db.connectObj.version_tag;
+  const locale = req.getLocale?.();
+  const scGlobals =
+    `<script>var _sc_globalCsrf = ${JSON.stringify(req.csrfToken())}` +
+    `, _sc_version_tag = ${JSON.stringify(version_tag)}` +
+    (locale ? `, _sc_locale = ${JSON.stringify(locale)}` : "") +
+    `, _sc_pageloadtag = Math.floor(Math.random() * 16777215).toString(16)` +
+    (req?.user?.role_id === 1 ? `, _sc_is_admin = true` : "") +
+    `;</script>`;
+  const normalized = html_string.replace(
+    /\/static_assets\/[a-f0-9]+\//g,
+    `/static_assets/${version_tag}/`
+  );
+  const html = normalized.includes("</head>")
+    ? normalized.replace("</head>", `${scGlobals}</head>`)
+    : scGlobals + normalized;
+  return res.send(html);
+};
+
+/**
  * Send HTML file to client without any menu
  * @param {any} req
  * @param {any} res
@@ -662,6 +691,7 @@ module.exports = {
   get_sys_info,
   admin_config_route,
   sendHtmlFile,
+  sendHtmlStringWithGlobals,
   setRole,
   getEligiblePage,
   getRandomPage,
